Clean up ApplicationController, mostly
